Make steps integer

+++ msc_root/steps/models.py
@@ -6,7 +6,7 @@ class StepEntry(models.Model):
     peaker = models.ForeignKey(User, on_delete=models.CASCADE)
     entered = models.DateTimeField(auto_now_add=True)
     date = models.DateField()
-    steps = models.FloatField()
+    steps = models.IntegerField()
     valid = models.BooleanField()
     notes = models.TextField()
